03-object-detection/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
albumentations>=1.3.0

# Metrics & Evaluation
torchmetrics>=1.0.0
scikit-learn>=1.3.0
pycocotools>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0.0
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional (Export & Optimization)
onnx>=1.14.0
onnxruntime>=1.15.0
# tensorrt (requires manual installation)
